Bump to 3.1.0
Bump the recipe to triton 3.1.0, using the commit referenced
in PyTorch's .ci/docker/ci_commit_pins/triton.txt (since upstream
did not tag the release).  Backport a few patches to make it compatible
with LLVM/MLIR 19, as the 3.1.0 release was made against a mailine
commit between LLVM 18 and 19, and does not work with either of these
versions.

TODO: unvendor nvidia components